refactor: start the refactor to use a React state to store the current state of the form
