+++ packages/twenty-front/src/modules/workflow/types/FunctionInput.ts
@@ -1,5 +1,3 @@
-export type FunctionInput =
-  | {
-      [name: string]: FunctionInput;
-    }
-  | any;
+export type FunctionInput = {
+  [key: string]: FunctionInput | string;
+};
